% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_dist.R
\name{area_dist}
\alias{area_dist}
\title{Area Distribution}
\usage{
area_dist(metadata)
}
\arguments{
\item{metadata}{"matrix" "array" ;  metadata of grdc dataset. Can be created by \link[dischanalyst]{metadata_grdc} function}
}
\value{
histogram , Distribution of Catchment Areas.
}
\description{
descriptive statistics. Function helps to gain overview over the distribution of the catchment area - sizes within grdc-dataset.
}
\examples{
\dontrun{ area_dist(metadata)}
}
