% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMQ.R
\name{dfMQ}
\alias{dfMQ}
\title{MQ Trend for multiple stations}
\usage{
dfMQ(data, seasonal = "Y", mod = 1)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{seasonal}{character; default:"Y". Possible Inputs:  "Y"( MQ Trend of (hydrological) Years); "WI"(MQ Trend of Winters during years); "SP"(MQ Trend of Springs during (hydrological) years); "SU"(MQ Trend of Summers during (hydrological) years); "AU"(MQ Trend of Autums during (hydrological) years)}

\item{mod}{numeric; possible input: 1,2,3. default value: 1; output of both: \link[zyp]{zyp.trend.vector}, \link[stats]{lm}. Defines the way to calculate intercept and slope. For mod=2  \link[zyp]{zyp.trend.vector} with PreWhitening by "yuepilon-method" is used. Sen-Slope-Approach used to define direction of the trend and the significance is  determined by Kendall's P-Value computed for the final detrendet time series. For mod=3: \link[stats]{lm} with a least squares approach is used.}
}
\value{
dataframe. Contains for every station columns: "Stationname", "intercept-zyp","slope-zyp","significance-zyp","linear model- intercept","linear model-slope","normalized slope", "Season" .
}
\description{
MQ Trend for multiple stations
}
\examples{
\dontrun{ dfMQ(data, seasonal="AU", 3)}

}
