% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdc_germany.R
\docType{data}
\name{grdc_germany}
\alias{grdc_germany}
\title{GRDC Germany- Datasets}
\format{
A data frame with 326 rows and 23 column(s)
\describe{
  \item{grdc_no}{Number given by GRDC to organize Data}
  \item{river}{Rivername as character}
  \item{station}{Stationname as character}
  \item{geometry}{spatial information of stations}

  includes other descriptive analytics
}
}
\source{
\url{https://www.bafg.de/GRDC/EN/Home/homepage_node.html}
}
\usage{
grdc_germany
}
\description{
This dataset helps to easily read your data of GRDC-Germany using . \code{grdc_readr} To use the function and this dataset, download data from:\url{https://www.bafg.de/GRDC/EN/Home/homepage_node.html}
}
\keyword{datasets}
