% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdc_list.R
\name{grdc_list}
\alias{grdc_list}
\title{GRDC List}
\usage{
grdc_list(metadata, path)
}
\arguments{
\item{metadata}{Data Frame. Overview of GRDC-Dataset.  The metadata can be created by \link[dischanalyst]{metadata_grdc} function}

\item{path}{character; pathway to grdc_discharge folder on computer}
}
\value{
Grdc List
}
\description{
Function returns List. Each entry contains Discharge Measurements as well as the corresponding date of a specific station. The Number of the stations depends on the metadata.It may take some time to load the GRDC-List.
}
\examples{
\dontrun{
grdc_list(metadata_germany, "/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/")
}


}
