% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdc_readr.R
\name{grdc_readr}
\alias{grdc_readr}
\title{GRDCC River Dataset Reader.}
\usage{
grdc_readr(data, rivername, path)
}
\arguments{
\item{data}{Dataframe ;  a Dataframe including columns: "river" (rivername as character), "station" (rivername as character) and "grdc_no" (grdc number as double)}

\item{rivername}{character ; name of the demanded river. Must equal name in column "river" in the Dataframe}

\item{path}{character; Pathway of your computer where the GRDC dataset is saved.}
}
\value{
dataset of the demanded river
}
\description{
GRDCC River Dataset Reader.
}
\examples{

\dontrun{
mosel=grdc_readr(data,"MOSELLE RIVER","/Users/username/Desktop/worfolder/datafolder/grdc_03_2021/grdc_disc/" )
}






}
