% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_distribution.R
\name{length_distribution}
\alias{length_distribution}
\title{Distribution of Length of Discharge Measurements of all Stations within Metadata}
\usage{
<<<<<<< HEAD
<<<<<<< HEAD
length_distribution(metadata)
=======
length_distribution(metadata, type = "dens")
>>>>>>> ffc5ade01805f6d97cd5bc4bbf8c7168d89c11ec
=======
length_distribution(metadata, type = "dens")
>>>>>>> ffc5ade01805f6d97cd5bc4bbf8c7168d89c11ec
}
\arguments{
\item{metadata}{"matrix" "array" ;  Metadata of grdc dataset. Can be created by \link[dischanalyst]{metadata_grdc} function}
}
\value{
density plot
}
\description{
Distribution of Length of Discharge Measurements of all Stations within Metadata
}
\examples{
\dontrun{
length_distribution(metadata)}

}
