% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_grdc.R
\name{metadata_grdc}
\alias{metadata_grdc}
\title{Create Metadata for GRDC Dataset}
\usage{
metadata_grdc(Country, path)
}
\arguments{
\item{Country}{character; abbrevation used in GRDC Dataset for specific country. e.g. "DE" for Germany.}

\item{path}{character; pathway to grdc_discharge folder on computer}
}
\value{

}
\description{
Create Metadata for GRDC Dataset
}
