% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_trend.R
\name{min_trend}
\alias{min_trend}
\title{Calculates minimum Value for every year since the begin of the measurements. Returns Linear Model Coefficients for the Minimum Values over the years. Minimum Values ~ Years.}
\usage{
min_trend(data, station, Name)
}
\arguments{
\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{Name}{character; Name of the Dataset. e.g. "Mosel"}
}
\value{

}
\description{
Calculates minimum Value for every year since the begin of the measurements. Returns Linear Model Coefficients for the Minimum Values over the years. Minimum Values ~ Years.
}
\examples{
\dontrun{ min_value_lm(mosel, "COCHEM", "Mosel")}
}
