% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintrendmeta.R
\name{mintrendmeta}
\alias{mintrendmeta}
\title{Area-wide Minimum-Values Trend}
\usage{
mintrendmeta(metadata, data, Startyear, Endyear)
}
\arguments{
\item{metadata}{data.frame; Overview of GRDC-Dataset.  The metadata can be created by \link[dischanalyst]{metadata_grdc} function.}

\item{data}{list; Contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{Startyear}{numeric; Startyear of timerange.}

\item{Endyear}{numeric; Endyear of timerange.}
}
\value{
dataframe. Including the stationname, the river, the spatial information of the station, the trend (linear model and zyp/"yuepilon"approach (with PreWhitening and Autocorrelation)) trend within every year/winter/summer/spring/autumn in timeframe.
\describe{
  \item{intercept_zyp}{intercept created by \link[zyp]{zyp.trend.vector}}
  \item{slope_zyp}{slope created by \link[zyp]{zyp.trend.vector}}
  \item{sig_zyp}{significance (Kendall's P-Value) for the final detrended time-series}
  \item{intercept_ls}{intercept created by \link[stats]{lm}}
  \item{slope_ls}{slope created by \link[stats]{lm}}
}
}
\description{
Calculating Trend for annual minimum values. It is  possible to calculate the trend for the  minimum values of a season or for the whole hydrological year.
}
