% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_trendset.R
\name{MQ_trendset}
\alias{MQ_trendset}
\title{MQ_trendset}
\usage{
MQ_trendset(metadata, data, seasonal)
}
\arguments{
\item{metadata}{Data Frame. Overview of GRDC-Dataset.  The metadata can be created by \link[dischanalyst]{metadata_grdc} function}

\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{seasonal}{character; default:"Y". Possible Inputs:  "Y"( MQ Trend of (hydrological) Years); "WI"(MQ Trend of Winters during years); "SP"(MQ Trend of Springs during (hydrological) years); "SU"(MQ Trend of Summers during (hydrological) years); "AU"(MQ Trend of Autums during (hydrological) years)}
}
\value{
list; containing annual Mean Values (or Summer/Winter/Autumn/Spring), normalized Mean Value and Years.
}
\description{
MQ_trendset
}
