% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMxQ_trend.R
\name{NMxQ_trend}
\alias{NMxQ_trend}
\title{NMxQ_trend}
\usage{
NMxQ_trend(data, station, x, graphic = T)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{x}{length of period (days). With decreasing Values for x , the influence of short-term anthropogenic influences increases. E.g. x=7, x=14, x=30}

\item{graphic}{logical; default=T. For graphic=T, function returns a graph and visualizes the trend.   For graphic=F, function returns the model coefficients. For the Sens Sloap Approach (using: \link[zyp]{zyp.trend.vector}) and for the linear model approach (using: \link[stats]{lm}).}
}
\value{
Graph/ List.
}
\description{
Function calculates for every year within measurements NMxQ Value (Description: \link[dischanalyst]{NMxQ}). It returns a trend: Sens Sloap and Linear Model.
}
\examples{
 \dontrun{
list=NMxQ_trend(data, "COCHEM", 4, graphic=F)}

}
