% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmxqmeta.R
\name{NMxQmeta}
\alias{NMxQmeta}
\title{NMxQmeta}
\usage{
NMxQmeta(x, data, metadata, Startyear, Endyear)
}
\arguments{
\item{x}{length of period (days). With decreasing Values for x , the influence of short-term anthropogenic influences increases. E.g. x=7, x=14, x=30}

\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{metadata}{Data Frame. Overview of GRDC-Dataset.  The metadata can be created by \link[dischanalyst]{metadata_grdc} function.}

\item{Startyear}{numeric; startyear of timerange.}

\item{Endyear}{numeric; endyear of timerange.}
}
\value{
dataframe. Including the stationname, the river, the spatial information of the station, the trend (linear model and zyp/"yuepilon"approach (with PreWhitening and Autocorrelation)) trend within every year/winter/summer/spring/autumn in timeframe.
}
\description{
Function creates dataframe of stations, containing Trend (every year/within a specific season) of their  NMxQ Value within the given timeframe. Therefore it filters all stations within the metadataset depending on the start- and endyear of their Measurement Series.
}
\examples{
 \dontrun{
NMxQmeta(7 , data, metadata, 1820, 2019)}

}
