% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_plot_time_series.R
\name{Q_plot}
\alias{Q_plot}
\title{Discharge time series plot}
\usage{
Q_plot(data, Name, station)
}
\arguments{
\item{data}{River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations}

\item{Name}{as character. Name of the River. e.g. "Mosel"}

\item{station}{as character. Name of the Station e.g. "COCHEM" - must be named the same like list entry in data.}
}
\value{
Graphic showing Discharge time series.
}
\description{
Desciptive Statistics. Time Series of Discharge at specific station.
}
