% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QBoxplot_year.R
\name{QBoxplot_year}
\alias{QBoxplot_year}
\title{Boxplot of Discharge Measurements in a specific year}
\usage{
QBoxplot_year(data, Name, station, year, h = T)
}
\arguments{
\item{data}{River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{Name}{as character. Name of the River. e.g. "Mosel"}

\item{station}{as character. Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{year}{as numeric. a certain year within the time series since begin of measurements}

\item{h}{hydrological year: TRUE or FALSE. If h=T, output is boxplot of hydrological year (given year/given year+1). If h=F output is Boxplot of Values of calendric year.}
}
\value{
Boxplot Graphic of Discharge Measurements of specific (hydrological/calendrical) year
}
\description{
Boxplot of Discharge Measurements in a specific year
}
\examples{
\dontrun{ Boxplot_year(mosel, "MOSELLE RIVER", "COCHEM", 2000,h=T)}


}
