% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qmin_trend.R
\name{Qmin_trend}
\alias{Qmin_trend}
\title{Trend of Annual Discharge Minima since begin of Measurements}
\usage{
Qmin_trend(data, station)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}
}
\value{

}
\description{
Trend of Annual Discharge Minima since begin of Measurements
}
\examples{
\dontrun{ Qmin_trend(mosel, "COCHEM")}

}
