% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qplot.R
\name{Qplot}
\alias{Qplot}
\title{Discharge time series plot}
\usage{
Qplot(data, Name, station)
}
\arguments{
\item{data}{River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{Name}{as character. Name of the River. e.g. "Mosel"}

\item{station}{as character. Name of the Station e.g. "COCHEM" - must be named the same like list entry in data.}
}
\value{
Graphic showing Discharge time series.
}
\description{
Desciptive Statistics. Time Series of Discharge at specific station. Including all measurements in list.
}
\examples{
\dontrun{ Q_plot(mosel, "Mosel", "COCHEM")}

}
