% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qploty.R
\name{Qploty}
\alias{Qploty}
\title{Discharge Plot in specific (calendrical/hydrological) year}
\usage{
Qploty(data, station, year, h = T)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{year}{numeric; a certain year within the time series since begin of measurements.}

\item{h}{logical; hydrological year. If h=TRUE; hydrological year November - October (given year/given year +1). If h=FALSE: calendrical year: January- December.}
}
\value{
graphic of discharge time series in given year
}
\description{
Discharge Plot in specific (calendrical/hydrological) year
}
\examples{
\dontrun{
Qploty(mosel, "COCHEM", 2000, h=T)
}


}
