% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantile_freq.R
\name{Quantile_freq}
\alias{Quantile_freq}
\title{Trend of Values under quantile based Threshold U}
\usage{
Quantile_freq(data, station, quantile, graphic = T)
}
\arguments{
\item{data}{list; Contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the station. Must equal entry in data list.}

\item{quantile}{numeric; If Input is 0.1; 90% of all Values are bigger than this Value.}

\item{graphic}{default=T; if T, returns Histogram and Densityplot. if F; returns vector containing trends, calculated with zyp/stats.}
}
\value{
Graph/ list
\describe{
  \item{intercept_zyp}{intercept created by \link[zyp]{zyp.trend.vector}}
  \item{slope_zyp}{slope created by \link[zyp]{zyp.trend.vector}}
  \item{sig_zyp}{significance (Kendall's P-Value) for the final detrended time-series}
  \item{intercept_ls}{intercept created by \link[stats]{lm}}
  \item{slope_ls}{slope created by \link[stats]{lm}}
}
}
\description{
Function enables user to see whether the frequency of daily averages under a quantile based threshold is increasing or decreasing.
}
\examples{
\dontrun{ Quantile_freq(data, "COCHEM", 0.1)}

}
