% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonpl.R
\name{seasonpl}
\alias{seasonpl}
\title{Seasonal plot}
\usage{
seasonpl(data, station, Startyear, Endyear, month_start, month_end)
}
\arguments{
\item{data}{list; River from GRDC - Dataset. Output of \link[dischanalyst]{grdc_readr}. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{station}{character- must equal name of station in data.}

\item{Startyear}{numeric; Startyear.}

\item{Endyear}{numeric; Endyear.}

\item{month_start}{character; Begin of the season. E.g. Febuary="02"}

\item{month_end}{character: End of the seasoan. E.g. August="11"}
}
\value{
seasonal plot.
}
\description{
Function returns a discharge plot of a given season at a given station.
It allows the user to compare the same season/timespan within different hydrological years.
}
\examples{
\dontrun{
seasonpl(mosel, "COCHEM", 2013, 2017, "02", "05")
}

}
