% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seasonplot.R
\name{seasonplot}
\alias{seasonplot}
\title{Seasonplot}
\usage{
seasonplot(
  data,
  station,
  month_start,
  month_end,
  hyear = F,
  Startyear,
  Endyear
)
}
\arguments{
\item{data}{list; River from GRDC - Dataset. Output of \link[dischanalyst]{grdc_readr}. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{station}{character- must equal name of station in data.}

\item{month_start}{character- which month of the year. E.g. Febuary= "02"}

\item{month_end}{character- which month of the year. E.g. Febuary= "02"}

\item{hyear}{to gain on overview of the whole year.}

\item{Startyear}{numeric.}

\item{Endyear}{numeric. Endyear-Startyear>=4.}
}
\value{
plot of Seasons
}
\description{
function returns plot of a specific time frame within one year.
By defining a Startyear and an Endyear it enables you to compare a few years (at least 4).
}
\examples{
\dontrun{
seasonplot(mosel, "COCHEM", hyear=T, Startyear=2010,Endyear= 2018  )
}
}
