% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries}
\alias{timeseries}
\title{Plot timeseries of measurements}
\usage{
timeseries(
  metadata,
  path,
  startyear,
  endyear,
  type = geom_smooth(method = "auto")
)
}
\arguments{
\item{metadata}{"matrix" "array" ;  metadata of grdc dataset. Can be created by metadata_grdc function}

\item{path}{character; pathway to grdc_discharge folder on computer}

\item{startyear}{integer; year- when did time series begin}

\item{endyear}{integer; year- when did time series end}

\item{type}{type of visualization in ggplot. default: geom_line() (e.g. geom_point(), geom_path()), geom_smooth(method="auto"))}
}
\value{

}
\description{
Plot timeseries of measurements
}
\examples{
\dontrun{
timeseries(metadata, "/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/" , 1990,2020)
}

}
