% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_period_sum.R
\name{U_period_sum}
\alias{U_period_sum}
\title{Sum of Days under U within all measurements}
\usage{
U_period_sum(U, station, data)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns sum of  days  Value U within all measurements.}

\item{station}{haracter; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}
}
\value{

}
\description{
Sum of Days under U within all measurements
}
\examples{
\dontrun{ U_period_sum(88, "TRIER UP", mosel)}
}
