% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_period_year.R
\name{U_period_year}
\alias{U_period_year}
\title{Max. undercut length of value U within a specific (calendrical/hydrological) year}
\usage{
U_period_year(U, station, data, year, h)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns longest time measured under Value U in specific year.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{year}{character; specific year}

\item{h}{logical;  hydrological year. If h=TRUE; hydrological year November - October (given year/given year +1). If h=FALSE: calendrical year: January- December.}
}
\value{

}
\description{
Max. undercut length of value U within a specific (calendrical/hydrological) year
}
\examples{
\dontrun{ U_period_year(80, "COCHEM", mosel, "2003", h=T)}

}
