% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_periodploty.R
\name{U_periodploty}
\alias{U_periodploty}
\title{U_periodploty}
\usage{
U_periodploty(data, station, U, year, graph = T)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{U}{numeric;  Limit Value.}

\item{year}{logical;  hydrological year November - October (given year/given year +1)}

\item{graph}{logical; default=T. If graph=T, function returns graph, visualizing LowFlowPeriod. If graph=F, function returns List containing characteristic Values of LowFlow Period.}
}
\value{
Graph/List. Visualizing/ characterizing Low Flow Period. Threshold based.
}
\description{
U_periodploty
}
\examples{
\dontrun{ list=periodplot_thres(data, "COCHEM", 150, 2009,F)}
}
