% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_prob.R
\name{U_prob}
\alias{U_prob}
\title{Probability to to receive a smaller Discharge Value than U}
\usage{
U_prob(U, station, data)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns possibility to receive Value smaller than U.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}
}
\value{

}
\description{
Probability to to receive a smaller Discharge Value than U
}
\examples{
\dontrun{ U_prob(88, "TRIER UP", mosel)}

}
