% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_probd.R
\name{U_probd}
\alias{U_probd}
\title{Probability to receive Discharge Values under U within the decades since begin of measurements}
\usage{
U_probd(U, station, data)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns possibility to receive Value smaller than U for every decade since begin of measurements.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{Name}{character; Name of the Dataset. e.g. "Mosel"}
}
\value{

}
\description{
Probability to receive Discharge Values under U within the decades since begin of measurements
}
\examples{
\dontrun{ U_probd(150,"COCHEM", mosel)}

}
