% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_proby.R
\name{U_proby}
\alias{U_proby}
\title{Probabilty to fall under U for every year since begin of the measurements}
\usage{
U_proby(U, station, data, h = T)
}
\arguments{
\item{U}{numeric; Limit Value. Function calculates probability to fall below U for every year since the begin of the measurements. Returns a Graphic and a Trend over the years.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{h}{logical;  hydrological year. If h=TRUE; hydrological year November - October (given year/given year +1). If h=FALSE: calendrical year: January- December.}
}
\value{

}
\description{
Probabilty to fall under U for every year since begin of the measurements
}
\examples{
\dontrun{ U_proby(150,"COCHEM", data,h=T)}

}
