% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uprob.R
\name{Uprob}
\alias{Uprob}
\title{Probability to to have a smaller Discharge Value than U}
\usage{
Uprob(U, station, data)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns possibility to receive Value smaller than U.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}
}
\value{

}
\description{
Probability to to have a smaller Discharge Value than U
}
\examples{
\dontrun{ U_prob(88, "TRIER UP", mosel)}

}
