% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usum_period.R
\name{Usum_period}
\alias{Usum_period}
\title{Sum of Days under U within all measurements}
\usage{
Usum_period(U, station, data)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns sum of  days  Value U within all measurements.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}
}
\value{

}
\description{
Sum of Days under U within all measurements
}
\examples{
\dontrun{ Usum_period(88, "TRIER UP", mosel)}
}
