% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usum_periody.R
\name{Usum_periody}
\alias{Usum_periody}
\title{Sum of Days under U within a specific (calendrical/hydrological) yeaf}
\usage{
Usum_periody(U, station, data, year, h)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns sum of days under given Value U in a year.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{h}{logical;  hydrological year. If h=TRUE; hydrological year November - October (given year/given year +1). If h=FALSE: calendrical year: January- December.}
}
\value{

}
\description{
Function returns sum of days under a certain value (at a given measuring point of a given river),
within one year.
}
\examples{
\dontrun{ Usum_periody(88, "TRIER UP", mosel, "2003", h=T)}

}
