% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdc_readr.R
\name{grdc_readr}
\alias{grdc_readr}
\title{GRDCC River Dataset Reader.}
\usage{
grdc_readr(metadata, rivername, path)
}
\arguments{
\item{metadata}{metadata "matrix" "array" ;  metadata of grdc dataset. Can be created by \link[dischanalyst]{metadata_grdc} function}

\item{rivername}{character; name of the demanded river. Must equal name in column "river" of metadata}

\item{path}{character; Pathway of your computer where the GRDC dataset is saved.}
}
\value{
dataset of the demanded river
}
\description{
GRDCC River Dataset Reader.
}
\examples{

\dontrun{
mosel=grdc_readr(data,"MOSELLE RIVER","/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/" )
}






}
