% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_grdc.R
\name{metadata_grdc}
\alias{metadata_grdc}
\title{Create Metadata for GRDC Dataset}
\usage{
metadata_grdc(Country, path)
}
\arguments{
\item{Country}{character; Abbrevation used in GRDC Dataset for specific country. e.g. "DE" for Germany.}

\item{path}{character; Pathway to local grdc_discharge folder on computer}
}
\value{
data.frame; metadata of GRDC -Dataset and a given Country. Metadata contains information about: Ther GRDC-number, the name of the river, the name of the station, the country, the catchment area, the altitude, the startmonth/startyear, the endmonth/endyear, the length of the timeseries, the longitude and the latitude.
}
\description{
Returns metadataset for one country of local folder containing grdc-dataset.
}
\examples{
\dontrun{
metadata_germany=metadata_grdc("DE","/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/" )
}
#' @source \url{https://www.bafg.de/GRDC/EN/Home/homepage_node.html}

}
