% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_grdc.R
\name{metadata_grdc}
\alias{metadata_grdc}
\title{Create Metadata for GRDC Dataset}
\usage{
metadata_grdc(Country, path)
}
\arguments{
\item{Country}{character; abbrevation used in GRDC Dataset for specific country. e.g. "DE" for Germany.}

\item{path}{character; pathway to grdc_discharge folder on computer}
}
\value{
data.frame; metadata of GRDC -Dataset and a given Country
}
\description{
Create Metadata for GRDC Dataset
}
\examples{
\dontrun{
metadata_germany=metadata_grdc("DE","/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/" )
}
#' @source \url{https://www.bafg.de/GRDC/EN/Home/homepage_node.html}

}
