% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_repg.R
\name{metadata_repg}
\alias{metadata_repg}
\title{Representative levels in Germany}
\usage{
metadata_repg(metadata)
}
\arguments{
\item{metadata}{"matrix" "array" ;  metadata of grdc dataset. Can be created by \link[dischanalyst]{metadata_grdc} function}
}
\value{
"matrix". Same Arrangement like metadata (input- parameter). Contains representative stations.
}
\description{
this fucntion only contains representative levels for Germany.
}
\examples{
\dontrun{metadata_repg(metadata_germany)}

}
