% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_trend.R
\name{min_trend}
\alias{min_trend}
\title{Trend of minimum value}
\usage{
min_trend(data, station, mod = 1)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the station. Must equal one entry of the data list.}

\item{mod}{numeric; possible input: 1,2,3. default value: 1; output of both: \link[zyp]{zyp.trend.vector}, \link[stats]{lm}. Defines the way to calculate intercept and slope. For mod=2  \link[zyp]{zyp.trend.vector} with PreWhitening by "yuepilon-method" is used. Sen-Slope-Approach used to define direction of the trend and the significance is  determined by Kendall's P-Value computed for the final detrendet time series. For mod=3: \link[stats]{lm} with a least squares approach is used.}
}
\value{
list
\describe{
  \item{intercept_zyp}{intercept created by \link[zyp]{zyp.trend.vector}}
  \item{slope_zyp}{slope created by \link[zyp]{zyp.trend.vector}}
  \item{sig_zyp}{significance (Kendall's P-Value) for the final detrended time-series}
  \item{intercept_ls}{intercept created by \link[stats]{lm}}
  \item{slope_ls}{slope created by \link[stats]{lm}}
}
}
\description{
Calculates minimum Value for every year since the begin of the measurements. Coefficiencts for a model. Uses least squares approach with a higher uncertainity and Sen-Sloap approach with "Yuepilon" PreWhitening.
}
\examples{
\dontrun{ min_trend(mosel, "TRIER UP")}
}
