% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QBoxplot_year.R
\name{QBoxplot_year}
\alias{QBoxplot_year}
\title{Discharge Boxplot of specific (year}
\usage{
QBoxplot_year(data, Name, station, year, h)
}
\arguments{
\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{Name}{character; Name of the River. e.g. "Mosel".}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{year}{numeric; year within time series of measurements.}

\item{h}{logical; hydrological year. If h=TRUE; hydrological year November - October (given year/given year +1). If h=FALSE: calendrical year: January- December.}
}
\value{
Boxplot graphic of Discharge Measurements in a specific (calendrical/hydrological) year.
}
\description{
Discharge Boxplot of specific (year
}
\examples{
\dontrun{ QBoxplot_year(mosel, "MOSEL", "COCHEM", 2000, h=T)}

}
