% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qmin_trend.R
\name{Qmin_trend}
\alias{Qmin_trend}
\title{Trend of Annual Discharge Minima since begin of Measurements}
\usage{
Qmin_trend(data, station, Name)
}
\arguments{
\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{Name}{character; Name of the River. e.g. "Mosel"}
}
\value{

}
\description{
Trend of Annual Discharge Minima since begin of Measurements
}
\examples{
\dontrun{ Qmin_trend(mosel, "COCHEM", "MOSEL")}

}
