% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qplot.R
\name{Qplot}
\alias{Qplot}
\title{Discharge Plot of the entire time series of measurements}
\usage{
Qplot(data, station)
}
\arguments{
\item{data}{list; River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data}

\item{Name}{character; Name of the River. e.g. "Mosel"}
}
\value{
Graphic showing Discharge time series.
}
\description{
Desciptive Statistics. Time Series of Discharge at specific station. Including all measurements in list.
}
\examples{
\dontrun{ Qplot(mosel, "COCHEM")}

}
