% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qplot.R
\name{Qplot}
\alias{Qplot}
\title{Discharge Plot of the entire time series of measurements}
\usage{
Qplot(data, station)
}
\arguments{
\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data}
}
\value{
Graphic showing Discharge time series.
}
\description{
Desciptive Statistics. Time Series of Discharge at specific station. Including all measurements in list.
}
\examples{
\dontrun{ Qplot(mosel, "COCHEM")}

}
