% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qplot_year.R
\name{Qplot_year}
\alias{Qplot_year}
\title{Discharge Plot in specific year}
\usage{
Qplot_year(data, Name, station, year)
}
\arguments{
\item{data}{River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}

\item{Name}{as character. Name of the River. e.g. "Mosel"}

\item{station}{as character. Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{year}{as numeric. a certain year within the time series since begin of measurements}
}
\value{
graphic of discharge time series in given year
}
\description{
Discharge Plot in specific year
}
\examples{
\dontrun{ Qplot_year(mosel, "Mosel", "COCHEM", 2000)}
{ Qplot_year(mosel, "MOSEL", "COCHEM", 2001)}
}
