% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantile_freq.R
\name{Quantile_freq}
\alias{Quantile_freq}
\title{Trend of Values under quantile based Threshold U}
\usage{
Quantile_freq(data, station, quantile, graphic = T)
}
\arguments{
\item{data}{list; Contains all stations that the discharge analysis should consider. List can be created by dischanalyst. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{station}{character; Name of the station. Must equal entry in data list.}

\item{quantile}{numeric; If Input is 0.1; 90% of all Values are bigger than this Value.}

\item{graphic}{default=T; if T, returns Histogram and Densityplot. if F; returns vector containing trends, calculated with zyp/stats.}
}
\value{

}
\description{
Trend of Values under quantile based Threshold U
}
\examples{

\dontrun{ Quantile_freq(data, "COCHEM", 0.1)}

}
