% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiseger.R
\name{tiseger}
\alias{tiseger}
\title{Plot timeseries of measurements}
\usage{
tiseger(metadata, path, startyear, endyear, type = geom_line())
}
\arguments{
\item{metadata}{"matrix" "array" ;  metadata of grdc dataset. Can be created by \link[dischanalyst]{metadata_grdc} function}

\item{path}{character; pathway to grdc_discharge folder on computer. E.g. "/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/"}

\item{startyear}{integer; last possible startyear of timeseries. Minimum of xlim for plot.}

\item{endyear}{integer; minimum endyear of timeseries. maximum of xlim for plot.}

\item{type}{type of visualization in ggplot. default: geom_line() (e.g. geom_point(), geom_path()), geom_smooth(method="auto"))}

\item{metadata_repg}{"matrix" "array" ;  metadata of grdc dataset germany only contains relevant levels for Germany. Can be created by \link[dischanalyst]{metadata_repg} function.}
}
\value{

}
\description{
Function plots timeseries within a certain timeframe.
To provide clarity the function reduces the amount of displayed timeseries if the number of stations with measurements >= the timeframe is above 20.
The timeseries are then reduced to the most representative timeseries within the timeframe. As the representative levels within \link[dischanalyst] were determined only for Germany,
this function only produces a meaningflu result for german stations
}
\examples{
\dontrun{
timeseriesg(metadata, "/Users/username/Desktop/folderone/datafolder/grdc_03_2021/grdc_disc/" , 1990,2020, metadata_repg= metadata_repg)
}

}
