% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_period.R
\name{U_period}
\alias{U_period}
\title{Max undercut length of value U within all measurements}
\usage{
U_period(U, station, data)
}
\arguments{
\item{U}{as numeric, Limit Value. Function returns longest time measured under Value U.}

\item{station}{as character. Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{River from GRDC - Dataset. Output of grdc-readr function. Type: list; list entries: measurement stations. For every Station: Date since begin of Measurements (as character) and Value (as numeric).}
}
\value{

}
\description{
Function indicates how long the longest period existed that a certain value was undercut (at a given measuring point of a given river),
within the time period of the measurements.
}
\examples{
\dontrun{ U_period(88, "TRIER UP", mosel)}
}
