% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_periody.R
\name{U_periody}
\alias{U_periody}
\title{Max. amount of days under Value U within a specific (calendrical/hydrological) year}
\usage{
U_periody(U, station, data, year, h)
}
\arguments{
\item{U}{numeric; Limit Value. Function returns longest time measured under Value U in specific year.}

\item{station}{character; Name of the Station e.g. "COCHEM" - must be named equally like list entry in data.}

\item{data}{list; contains all stations that the discharge analysis should consider. List can be created by \link[dischanalyst]{grdc_list}. Each entry of the list contains the existing discharge measurements (as numeric) and the corresponding dates (as character) for the station.}

\item{h}{logical;  hydrological year. If h=TRUE; hydrological year November - October (given year/given year +1). If h=FALSE: calendrical year: January- December.}
}
\value{

}
\description{
Function indicates how long the longest period existed that a certain value was undercut (at a given measuring point of a given river within a given year),
within the time period of the measurements.
}
\examples{
\dontrun{ U_periody(80, "COCHEM", data, "2003", h=T)}

}
